Safari-609.1.11.1/LayoutTests/fast/forms/file/entries-api/drag-folder-webkitEntries-expected.txt

Tests dragging a folder onto a file input without webkitdirectory set and querying input.webkitEntries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputElement.webkitdirectory is false
PASS entries.length is 2
PASS entries[0].name is "test.txt"
PASS entries[0].fullPath is "/test.txt"
PASS entries[0].isFile is true
PASS entries[1].name is "testFiles"
PASS entries[1].fullPath is "/testFiles"
PASS entries[1].isDirectory is true
PASS files.length is 2
PASS files[0].name is "test.txt"
PASS files[0].size is 5
PASS files[0].type is "text/plain"
PASS files[1].name is "testFiles"
PASS files[1].size is one of `272`, and `256`
PASS files[1].type is ""
PASS successfullyParsed is true

TEST COMPLETE
